feat: symbolic link deployment

+++ src/package/mod.rs
@@ -12,7 +12,7 @@ use crate::{
     cli::{DeployArgs, ImportArgs, UpdateArgs},
     context::Context,
     utils::{
-        BACKUP_EXT, LogLevel, cprintln, get_string_from_value, get_string_hashmap_from_value,
+        self, BACKUP_EXT, LogLevel, cprintln, get_string_from_value, get_string_hashmap_from_value,
         get_vec_string_from_value, normalize_home_path, resolve_path, string_hashmap_to_toml_table,
         vec_string_to_toml_array,
     },
@@ -41,6 +41,8 @@ pub struct Package {
     pub prompts: HashMap<String, String>, // Package-level prompts
     #[serde(default)]
     pub ignore: Vec<String>, // Patterns to ignore during deployment
+    // Symlinks defaults to false
+    pub symlink: bool,
 }
 
 #[derive(Eq, Hash, PartialEq, Debug, Clone, Copy)]
@@ -64,6 +66,7 @@ impl Package {
             skip: false,
             prompts: HashMap::new(),
             ignore: Vec::new(),
+            symlink: false,
         }
     }
 
@@ -116,6 +119,10 @@ impl Package {
             .expect("The 'prompts' field must be a table of string to string mappings");
         let ignore = get_vec_string_from_value(pkg_val.get("ignore"))
             .expect("The 'ignore' field must be an array of strings");
+        let symlink = pkg_val
+            .get("symlink")
+            .and_then(|v| v.as_bool())
+            .unwrap_or(false);
 
         Ok(Self {
             name: pkg_name.to_string(),
@@ -129,6 +136,7 @@ impl Package {
             targets,
             prompts,
             ignore,
+            symlink,
         })
     }
 
@@ -175,6 +183,7 @@ impl Package {
         if !self.ignore.is_empty() {
             pkg_table.insert("ignore".to_string(), vec_string_to_toml_array(&self.ignore));
         }
+        pkg_table.insert("symlink".to_string(), toml::Value::Boolean(self.symlink));
         pkg_table
     }
 
@@ -238,6 +247,16 @@ impl Package {
 
     pub fn should_ignore(&self, rel_path: &Path) -> bool {
         should_ignore(&self.ignore, rel_path)
+    }
+
+    pub fn resolve_deployment_path(&self, ctx: &Context) -> anyhow::Result<PathBuf> {
+        if !self.symlink {
+            anyhow::bail!("Package '{}' is not configured for symlinking", self.name);
+        }
+        Ok(ctx
+            .working_dir
+            .join(utils::SYMLINK_FOLDER)
+            .join(&self.name))
     }
 
     /// Backup the package by copying files from dest to a backup location, recursively.
@@ -405,7 +424,7 @@ impl Package {
                 return Ok(BackupDeployResult::Skipped);
             }
             if !dry_run {
-                if backup && dest.exists() {
+                if backup && dest.exists() && !self.symlink {
                     let backup_path = create_backup_path(dest);
                     std::fs::copy(dest, &backup_path)?;
                 }
@@ -414,7 +433,7 @@ impl Package {
         } else {
             // It can be a binary file, copy as-is and return Ok
             if !dry_run {
-                if backup && dest.exists() {
+                if backup && dest.exists() && !self.symlink {
                     let backup_path = create_backup_path(dest);
                     std::fs::copy(dest, &backup_path)?;
                 }
@@ -432,7 +451,11 @@ impl Package {
     ) -> Result<BackupDeployResult, anyhow::Error> {
         self.execute_pre_actions(ctx, args.dry_run)?;
         let copy_from = resolve_path(&self.src, &ctx.working_dir);
-        let copy_to = self.resolve_dest(ctx);
+        let copy_to = if self.symlink {
+            self.resolve_deployment_path(ctx)?
+        } else {
+            self.resolve_dest(ctx)
+        };
         let mut result = BackupDeployResult::Skipped;
         if copy_from.is_dir() {
             let mut all_deployed_paths = vec![];
@@ -470,6 +493,15 @@ impl Package {
             if let BackupDeployResult::Success = dep_result {
                 result = BackupDeployResult::Success;
                 println!("=> Deployed {}", copy_to.display());
+            }
+        }
+        if self.symlink {
+            let symlink_to = self.resolve_dest(ctx);
+            if !args.dry_run {
+                if symlink_to.exists() {
+                    std::fs::remove_file(&symlink_to)?;
+                }
+                std::os::unix::fs::symlink(&copy_to, &symlink_to)?;
             }
         }
         cprintln(
